SU-571 [User] I want to recover a forgotten password (#127)
# Description
New functionality that allows a user to reset a forgotten password. In the login dialog, the user is able to request to reset its StageUp password. A new dialog opens to request the email associated to the stageUp account and sends an email with the One-Time-Password (OTP). The reset password form opens from the link provided in the email and is only accessed if contains the OTP.  when the user clicks reset password, sends OTP & new password to backend, backend verifies OTP is valid & sets the password of the users account whose e-mail is associated with the OTP.

# New Features
1. New interface 'IPasswordReset' that describes the contents of the schema.
2. New Entity 'PasswordReset' for storing OTP that is associated with a user e-mail address.
3. Add link to sign-in form that opens forgotten password reset component.
4. New route/controller/service/component to request the recovery of a forgotten password.
5. Response email with link to reset password.
6. New route/controller/service/component with form to reset password.
7. Response confirmation email of the changes.

# Video of the Process
https://user-images.githubusercontent.com/72453357/114029233-3736d980-9871-11eb-8232-3b4eca3bf49d.mp4

# Emails Generated
![reset email](https://user-images.githubusercontent.com/72453357/114033216-f771f100-9874-11eb-8436-06cc0f6ba47e.PNG)

![confirm email](https://user-images.githubusercontent.com/72453357/114031945-c2b16a00-9873-11eb-82de-825287b4a6d0.PNG)

![expired](https://user-images.githubusercontent.com/72453357/114040078-2f7c3280-987b-11eb-97b6-58e9468b8c46.PNG)