SU-470 [Internal] Legal basis DB table & fields (#224)
- Added single table/entity `person-consent` that uses single table inheritance and a __type__ column to identify the type.
- This has relations to hosts, users, performances and also to documents in the  __consent__ table.
- Added interfaces/types for the different types of person-consent relationships (`IPersonConsent` which is a generic umbrella interface, `HostMarketingConsent`, `SuMarketingConsent`, `UserCookieConsent` and `PerformanceUploadConsent`).

<img width="1565" alt="Screenshot 2021-07-19 at 15 54 02" src="https://user-images.githubusercontent.com/48356897/126171689-bda7da9a-b5d8-4429-a6c1-5c6184a0245c.png">

I am really not sure if I have implemented this correctly. I am welcome to all feedback and can change any (or all!) if need be. Another thought process I had was to use a `primary_document` and `secondary_document` column (since the most amount of documents associated with any particular type of consent is two). Or even an array. However this might raise confusion with regards to the specific type of documents being referred to.