SU-118  [Host Owner] I want to delete my performance before its premier date (#235)
Merge of dependency injection task which rewrites the back end to use services. This was made off SU-118 (SU-118-di) (see https://github.com/StageUp/core/pull/254) . There are still some issues with SU-118 but it's important to get the DI implemented so that dev work can continue. A fix branch will be made for the original SU-118.