feature(SU-472) [User] I want to be reminded of StageUp's legitimate interests before StageUp passes my personal data onto third parties when I purchase a ticket (#266)
Displays a message to users informing them that if they buy tickets then they are agreeing to T&Cs and Privacy Policy. Also fixes a few small bugs.

Adds HTML and links to performance brochure (to latest T&Cs and Privacy Policies).
Implied consent given by the action of the user buying a ticket.
Fixed bug in profile-settings.component that was trying to load HTML variables before the component had been initialised.
Updated OptOutDialogComponent to send the enum value rather than key to backend.
Fixed problem with opt-out 'reasons' not being sent in emails properly. Added opt out reasons to backend i18n-tokens and yaml. I considered instead to simply send the translated reason as a 'string' to the backend and inserting it directly (since it is translated through a pipe on the frontend). However this would mean we can't carry out server side validation of the sent string. However it might not be a big issue and would perhaps simplify the code. I am happy to change it back if people think that would be better.