feature(SU-932) I want to see a carousel for scheduled events on the host's profile page  (#261)
Adds a carousel to the host profile page which contains the upcoming/scheduled events filtered by the host whose page you are on. The events are also ordered from earliest to latest dates.

Original Requirements:

Create a carousel on the host profile page which looks like the carousels on the feed page
Make the performances be filtered by the host and order them by earliest to latest.
Changes related to the task:

Modified the readFeed endpoint in MyselfController so that it can accept a host id and can return the upcoming performances from that host. This change is also reflected in the validation.
The front-end FeedService needed to be to align with the change above.
I created a CarouselBaseComponent which contains the common fields and functions needed for components which have a carousel like the feed and the host profile components.
I added the carousel itself to the host profile component template and modified the css (see note below).
Note: I think there might be some underlying issue with the carousel styling as if there are not enough elements in the carousel or the screen is too big or small the shape of the carousel component gets distorted.

Testing:

I am working on the tests now. I would like to debug something although I cannot make things be written to the console. I will read about that or if you have a tip I'd be glad to hear it. I also noticed that the tests depend on the state of the dev db, should the db be erased before testing?