feature(SU-903) [Host Owner]  I should not be able to cancel/delete my performance during its occurrence (#269)
Disabled the ability for performances to be cancelled/ deleted while live.

UPDATE As per Shreya's comment below, a check is now done to see if a vod performance is currently 'live', that is, it's currently within the performance's publicity period. To do this a check was made in the backend and all performances now have a PerformanceType associated with them in the Performance Table.

Changes:

Disables the cancel button if a performance's status is currently live (delete was already disabled)
Adds a material tooltip with a message informing the user why they cannot delete/cancel (message is localised)
N.B I had already implemented a check for live within the controller logic for the delete/ cancel actions, an error will be thrown if currently live

In the video below, the performance status is changed in the database directly because this would only change to live when mux detects a live stream and it is easier to demo the functionality by flipping a value.