feature(SU-959) Take the user to the brochure page when they click on a performance's link from a browser  (#286)
merge: feature(SU-959) Take the user to the brochure page when they click on a performance's link from a browser

I created a new page by reusing the performance brochure components. This way the link can be shared and it will lead to a separate page.

Additionally, Open Graph meta tags have been added to the same page to display link preview information. It doesn't work yet because of its dynamic nature, this meta information is added to the html later, so that on the social media sites this info isn't picked up. The plan is to revisit this topic in frames of another issue.